content-security-policy/script-src/script-src-strict_dynamic_double_policy_honor_whitelist.sub.html

<!DOCTYPE HTML>
<html>

<head>
    <title>Whitelists in a separate policy are honored with `strict-dynamic` in the script-src directive.</title>
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>

    <!-- CSP served:
    1) Content-Security-Policy: script-src 'strict-dynamic' 'nonce-dummy'
    2) Content-Security-Policy: script-src 'self' 'nonce-dummy'
    -->
</head>

<body>
    <h1>Whitelists in a separate policy are honored with `strict-dynamic` in the script-src directive.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        async_test(function(t) {
            var handler = function(e) {
                document.removeEventListener('securitypolicyviolation', handler);
                t.unreached_func('No CSP violation report has fired.');
            };
            document.addEventListener('securitypolicyviolation', handler);

            var e = document.createElement('script');
            e.id = 'whitelisted-appendChild';
            e.src = 'simpleSourcedScript.js';
            e.onload = t.step_func_done(function() {
                assert_equals(document.getElementById('whitelisted-appendChild').getAttribute('executed'), '1');
            });
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
        }, 'Script injected via `appendChild` is allowed with `strict-dynamic` + a nonce+whitelist double policy.');
    </script>

    <!-- This test must be synchronous because we have to wait until the old event handler is removed. -->
    <script nonce='dummy'>
        test(function(t) {
            var handler = function(e) {
                assert_equals(document.getElementById('nonWhitelisted-appendChild').getAttribute('executed'), null);
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                assert_equals(e.originalPolicy.trim(), "script-src 'self' 'nonce-dummy'");
                document.removeEventListener('securitypolicyviolation', handler);
                t.done();
            };
            document.addEventListener('securitypolicyviolation', handler);

            var e = document.createElement('script');
            e.id = 'nonWhitelisted-appendChild';
            e.src = '{{location[scheme]}}://{{domains[www2]}}:{{ports[http][0]}}/nonexisting.js';
            e.onload = t.unreached_func('OnLoad should not be triggered.');
            document.body.appendChild(e);
        }, 'Non-whitelisted script injected via `appendChild` is not allowed with `strict-dynamic` + a nonce+whitelist double policy.');
    </script>
</body>

</html>
